% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVH.R
\docType{data}
\name{DVH}
\alias{DVH}
\title{DVH data for 365 patients}
\format{A list with each element containing a dataframe, the first column is the dose the second column is the volume.}
\source{
\url{http://www.milanHosp.info/}
}
\usage{
DVH
}
\description{
The dataset contains dose and volumes of the rectum irradiated of 365 patients.
}
\keyword{datasets}
