% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractionation.R
\docType{data}
\name{fractionation}
\alias{fractionation}
\title{fractionation dose  for 365 patients}
\format{A vector of fractionation dose of each of the 365 patients.}
\usage{
fractionation
}
\description{
fractionation dose  for 365 patients
}
\keyword{datasets}
