% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntcp.R
\name{ntcp}
\alias{ntcp}
\title{Fits NTCP models from an ntcpObject object}
\usage{
ntcp(ntcpObject, model = c("LymanKB.EUD", "LymanKB.EUD.EQD2", "LogitKB.EUD",
  "LogitKB.EUD.EQD2"), nstart = 5)
}
\arguments{
\item{ntcpObject}{an ntcp object created by the function createNtcpObject.}

\item{model}{the type model to fit it should be one of: LymanEUD, LymanEUDEQD2, LogitEUD
and LogitEUDEQD2. The default is LymanEUD.}

\item{nstart}{the number of starting values from which the fitting engine will use as
the starting. If the defualt is nstart=5.}
}
\description{
ntcp returns a model object with fitted parameter values.
This function takes in an ntcpObject object, and can fit four different NTCP models.
}
