% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicity.R
\docType{data}
\name{toxicity}
\alias{toxicity}
\title{Binary toxicity values for 365 patients}
\format{A vector of 0 (no toxicity) and 1 (toxicity)  of each of the 365 patients.}
\usage{
toxicity
}
\description{
Binary toxicity values for 365 patients
}
\keyword{datasets}
